// src/lib/properties/types/properties.ts
export interface PropertyDB {
  access_instructions: string;
  email: string;
  address: string;
  phone_number: string;
  property_address: string;
  name: string;
  supervisor_name: string;
};

export interface PropertyDetail {
  property_name: string;
  property_address: string;
  supervisor_name: string;
  email: string;
  phone: string;
  access_instructions: string;
};